Created simple AI, and flowing game state
Game state flows between player turn and enemy turn. On enemy turn the artificial "intelligence" selects a random cell on the grid to hit. This will be expanded upon so it actually seems like its thinking e.g. if it hits a cell it will try and hit the cell next to it on next turn etc.