Implemented early version of AI
Easy AI:  Working. Simply just hits random squares on the board, should be hard to lose against.

Medium AI: Not working. Implements the "hunt" algorithm. has 3 states, random (same AI as easy), searching, attacking. Searching is initiated when the AI finds a battleship through random hits, will get 4 adjacent cells of hit cell and hit those 4 until it finds another battleship cell, it now knows the direction of the ship due to the the adjacent cell of the first battleship cell also being a battleship. Goes into attacking state and keeps hitting cells in that direction. When it hits a non battleship cell it SHOULD go back to random state but this isnt working atm.

Hard AI: Not implemented. Will implement this if I have time, there are some interesting and complex algorithms but will see

File is messy and lots of potential for refactoring but just focusing on getting something to work first.