Fixed previous commit
Okay previous unoptimized code did not actually work, the entire application crashed...

but... now it is fixed. I think. I know. I think I know. I'm just about content with the medium AI now, it can win in around 40 turns, there's still that element of luck for the AI with hunt mode where it just goes for random cells, but for medium difficulty this is fine.

can I be arsed to implement the probability algorithm for hard difficulty? No.